[VS-7340] Startup class combined with Program
